rag/backend/requirements.txt
openai
python-dotenv
flask
azure-core
azure-search-documents
flask-cors